srcs/user-service/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
psycopg2-binary==2.9.9
python-decouple==3.8
django-cors-headers==4.3.1
pytest==8.3.4
pytest-django==4.9.0
